+++ gitmap/constants/constants.go
@@ -13,7 +13,7 @@
 package constants
 
 // Version.
-const Version = "5.44.0"
+const Version = "5.45.0"
 
 // RepoPath is set at build time via -ldflags.
 var RepoPath = ""